+++ ReleaseNotes.md
@@ -9,6 +9,7 @@
 * `DictionaryUtility` key types must be `notnull`.
 * Add `Stream.ReadAsync`/`WriteAsync` overloads for `Memory<byte>` on .NET Standard 2.1 and .NET Core.
 * Obsolete `DictionaryUtility.CreateKeyValuePair` on .NET Standard 2.1 and .NET Core.
+* Obsolete `EnumerableUtility.EnumerateBatches` on .NET 6.
 
 ## 0.11.2